[matsys_controls] VMT and VTF preview panels should continue to work after ~49.7 days

+++ public/matsys_controls/vmtpreviewpanel.h
@@ -73,7 +73,8 @@ private:
 	Color m_LightColor;
 	float m_flLightIntensity;
 	Vector m_vecCameraDirection;
-	float m_flLastRotationTime;
+	// dimhotepus: Float -> double.
+	double m_flLastRotationTime;
 	bool m_bDrawIn3DMode;
 };
